--- backend/src/main/java/com/geekbrains/practice/services/IAuthService.java
@@ -1,5 +1,0 @@
-package com.geekbrains.practice.services;
-
-public interface IAuthService {
-  void startListening();
-}
